ft: update left column on pinned entity type tabs
